fixed entities relations and constraints

+++ measurement-storage-service/src/main/java/com/bpm/measurementstorageservice/rabbit/dto/SensorMeasurementEvent.java
@@ -2,6 +2,7 @@ package com.bpm.measurementstorageservice.rabbit.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -9,9 +10,9 @@ public record SensorMeasurementEvent(
         String sensorType,
         String sensorId,
         String location,
-        Double temperature,
-        Double humidity,
-        Double dewPoint,
+        BigDecimal temperature,
+        BigDecimal humidity,
+        BigDecimal dewPoint,
         LocalDateTime timestamp
 ) {
 }
